introduce working DELETE
ecspcs-180

One note with this commit is that an error is being thrown because the schema has been changed, but this branch
doesn't have updated code to deal with the changes. However it successfully deletes objects from the db, just
doesn't display what was deleted to the user.

+++ src/main/groovy/edu/oregonstate/mist/videogamesampleapi/resources/PlatformResource.groovy
@@ -93,7 +93,15 @@ class PlatformResource extends Resource {
     @DELETE
     @Produces(MediaType.APPLICATION_JSON)
     public Response deletePlatform(@PathParam("id") Integer id) {
-        platformDAO.deleteById(id)
-        Response.ok().build()
+        Platform returnPlatform = platformDAO.getPlatformById(id)
+        ResponseBuilder responseBuilder
+
+        if (returnPlatform == null) {
+            responseBuilder = notFound()
+        }
+        else {
+            platformDAO.deleteById(id)
+            responseBuilder = ok(returnPlatform)
+        }
     }
 }
